feat: Added tests for planet controller
